% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbalanced_EM.R
\name{alpha_cond_params}
\alias{alpha_cond_params}
\title{Compute conditional distribution of sires effect}
\usage{
alpha_cond_params(y_data, Sigma_E, Sigma_A, mu = rep(0, nrow(Sigma_E)))
}
\arguments{
\item{y_data}{Observed data that inherits "\code{fullsibdata}"}

\item{Sigma_E}{Between-individuals covariance matrix}

\item{Sigma_A}{Between-sires covariance matrix}

\item{mu}{Global mean}
}
\value{
A list of multivariate Gaussian parameters for the conditional
distribution of the sires random effect. Each entry is a list with entries
\code{mean} and \code{cov}, which are the mean vector and covariance matrix respectively.
}
\description{
Given the model
\deqn{
y_{ij} | \alpha_i \sim N(\alpha_i, \Sigma_E)
\alpha_i \sim N(0, \Sigma_A),
}{
y[ij] | alpha[i] ~ N(alpha[i], Sigma[E])
alpha[i] ~ N(0, Sigma[A]),
}
computes the parameters of the Gaussian conditional distribution
\deqn{
\alpha_i | y_i.
}{
alpha[i] | y[i].
}
}
