% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_fullsib.R
\name{balance_data}
\alias{balance_data}
\title{Compute expected data conditional on unbalanced data}
\usage{
balance_data(y_data, params)
}
\arguments{
\item{y_data}{An object inheriting "\code{fullsibdata}" containing unbalanced data}

\item{params}{A list of objects containing a field called \code{means} used to fill
in the unobserved data of \code{y_data}}
}
\value{
A balanced \code{fullsibdata} object.
}
\description{
Compute expected data conditional on unbalanced data
}
