% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_norm_reml.R
\name{block_conjugate}
\alias{block_conjugate}
\title{Conjugate each block of cond_cov by a matrix}
\usage{
block_conjugate(cond_cov, conj_mat, one_side = TRUE)
}
\description{
Conjugate each block of cond_cov by a matrix
}
