% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_norm.R
\name{cond_cov}
\alias{cond_cov}
\title{Compute the conditional covariances of two-way MANOVA random effects}
\usage{
cond_cov(init_covs, data, flat_sire = FALSE)
}
\arguments{
\item{init_covs}{A list of prior covariance matrices with entries named
\code{ind}, \code{dam} and \code{sire} indicating the between-individuals, between-dams
and between-sires covariances,}

\item{data}{An object inheriting \code{halfsibdata}}

\item{flat_sire}{A logical indicating whether a flat prior should be used
for the sire effect.}
}
\value{
A closure that takes the following arguments:
\itemize{
\item \code{i}: The name of sire
\item \code{j}, \code{k}: The name of dam, or \code{"group"}
}

The closure function returns the posterior covariance between the random
effects \code{beta[ij]} and \code{beta[ik]}. If \code{j == "group"}, returns the posterior
covariance of \code{alpha[i]} and \code{beta[ik]}. and similarly if \code{j == "group"}.
}
\description{
Under the model \code{y[ijk] == mu + a[i] + beta[ij] + epsilon[ijk]}, where
each \code{alpha[i]}, \code{beta[ij]} and \code{epsilon[ijk]} are independent mean-0,
\code{q}-dimensional normal random vectors with with covariance matrices
\code{Sigma[A]}, \code{Sigma[B]} and \code{Sigma[E]} respectively, compute the covariance
matrices of \verb{(alpha[i], beta[i1], ..., beta[iJ])} conditional on the
observed data for each \code{i}.
}
