% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbalanced_EM.R
\name{eigensqrt}
\alias{eigensqrt}
\title{Use eigen to compute a square root of the non-negative definite symmetric matrix A}
\usage{
eigensqrt(A)
}
\description{
The resulting U satisfies t(U) \%*\% U == A
}
