% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_halfsib.R
\name{extend_names}
\alias{extend_names}
\title{Create new names for dams based on their sires}
\usage{
extend_names(sires, existing_dams = character(0))
}
\description{
Create new names for dams based on their sires
}
