% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halfsibdesign.r
\docType{package}
\name{halfsibdesign}
\alias{halfsibdesign}
\title{\code{halfsibdesign} package}
\description{
An R package containing several algorithms for finding REML estimates for
covariances in the balanced q-dimensional half-sib design.
}
