% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_norm.R
\name{make_W}
\alias{make_W}
\title{Compute (Sigma\link{A} + sum((Sigma\link{B} + Sigma\link{E}/n\link{j})^(-1)))^(-1) for
values of n\link{j} given in \code{data}}
\usage{
make_W(init_covs, data, flat_sire = FALSE)
}
\description{
Compute (Sigma\link{A} + sum((Sigma\link{B} + Sigma\link{E}/n\link{j})^(-1)))^(-1) for
values of n\link{j} given in \code{data}
}
