% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manova.r
\name{manova}
\alias{manova}
\alias{manova_1way}
\alias{manova_2way}
\alias{manova_2way_df}
\title{Compute 1- or 2-way MANOVA estimates}
\usage{
manova_1way(Mw, r, Mb, n)

manova_2way(M1, I1, M2, I2, M3, I3)

manova_2way_df(df, ...)
}
\arguments{
\item{Mw, Mb}{Sum-of-squares within and between matrices}

\item{r, n}{Number of}

\item{df}{Data frame for 2-way design}
}
\description{
Given either a tuple of sum-of-squares matrices or a dataframe, computes
MANOVA estimates for the covariances
}
