% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manova.r
\name{manova_1way}
\alias{manova_1way}
\title{Compute 1- or 2-way MANOVA estimates}
\usage{
manova_1way(Mw, r, Mb, n)
}
\description{
Given either a tuple of sum-of-squares matrices or a dataframe, computes
MANOVA estimates for the covariances
}
