% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_fullsib.R
\name{naive_balance}
\alias{naive_balance}
\title{Balance the data by filling missing values with unconditional means}
\usage{
naive_balance(y_data)
}
\description{
This function is intended to provide initialisation values to the
EM algorithm
}
