% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reml.r
\name{reml_crit}
\alias{reml_crit}
\title{Compute the REML criterion}
\usage{
reml_crit(M1, S1, I1, M2, S2, I2, M3, S3, I3)
}
\arguments{
\item{M1, M2, M3}{Sum-of-squares matrices}

\item{S1, S2, S3}{Covariance estimates}

\item{I1, I2, I3}{Numbers of groups}
}
\description{
Computes the REML criterion for a 2-way design.
}
