% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfullsib.R
\name{rfullsib}
\alias{rfullsib}
\title{Simulate a dataset from a full-sib design}
\usage{
rfullsib(mu, A, I, E, J)
}
\arguments{
\item{mu}{Mean vector.}

\item{A, E}{Covariance matrices of the sires dams and individuals.}

\item{I, J}{Number of sires, dams per sire and individuals per dam.}
}
\description{
Simulate a dataset from a full-sib design
}
