% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhalfsib.r
\name{rhalfsib}
\alias{rhalfsib}
\title{Simulate a dataset from a half-sib design}
\usage{
rhalfsib(mu, A, I, B, J, E, K)
}
\arguments{
\item{mu}{Mean vector.}

\item{A, B, E}{Covariance matrices of the sires dams and individuals.}

\item{I, J, K}{Number of sires, dams per sire and individuals per dam.}
}
\value{
A \code{tibble} with column names \code{trait}, \code{value}, \code{sire},
\code{dam}, \code{individual}. The \code{value} column contains the response values whose
dimension is indexed by \code{trait}.
}
\description{
Sample responses from the following model:
\deqn{y_{ijk} = \mu + \alpha_i + \beta_{ij} + \epsilon_{ijk},}
where \eqn{\mu} is a deterministic mean vector and each \eqn{\alpha_i,
\beta_{ij}, \epsilon_{ijk}} are independent mean-zero multivariate
Gaussians with covariance matrices \eqn{A, B, E} respectively.
}
