% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_mats.r
\name{ss_mats}
\alias{ss_mats}
\title{Compute sum-of-squares matrices from half-sib design data}
\usage{
ss_mats(df)
}
\arguments{
\item{df}{Half-sib data frame with column names \code{trait}, \code{value}, \code{sire},
\code{dam}, \code{individual}}
}
\value{
A list with the following entries:
\itemize{
\item \code{m_ind}, \code{m_dam}, \code{m_sire}: The matrices S_E, S_B and S_A respectively
\item \code{q}: The dimensionality of the responses
\item \code{I}, \code{J}, \code{K}: The number of sires, dams per sire and individuals per dam
}
}
\description{
Given a dataframe of half-sib responses \eqn{y_{ijk}}, computes the following
matrices:
\deqn{S_A = \frac{JK}{I-1}\sum_{i} (\bar{y}_{i} - \bar{y}) (y_{i} - \bar{y})'}
\deqn{S_B = \frac{K}{I(J-1)}\sum_{i,j} (\bar{y}_{ij} - \bar{y}_{i})
(\bar{y}_{ij} - \bar{y}_{i})'}
\deqn{S_E = \frac{1}{IJ(K-1)}\sum_{i,j,k} (y_{ijk} - \bar{y}_{ij})
(y_{ijk} - \bar{y}_{ij})'}
These matrices satisfy
\deqn{E S_E = E,
E S_B = E + K B,
E S_A = E + K B + JK A,}
where A, B, E are the population covariance matrices of the sires, dams and
individuals respectively.
}
