% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepreml.r
\name{stepreml}
\alias{stepreml}
\alias{stepreml_1way}
\alias{stepreml_2way_mat}
\alias{stepreml_2way_df}
\title{Compute 1- or 2-way REML estimates with a stepwise method}
\usage{
stepreml_1way(A1, n1, A2, n2)

stepreml_2way_mat(
  M1,
  I1,
  M2,
  I2,
  M3,
  I3,
  max_iter = 50,
  err.tol = 1e-06,
  verbose = FALSE
)

stepreml_2way_df(df, ...)
}
\arguments{
\item{A1, A2}{Sum-of-squares matrices}

\item{n1, n2}{number of lines}

\item{df}{Data frame for 2-way design}
}
\description{
Given either a tuple of sum-of-squares matrices or a dataframe, computes
stepwise REML estimates for the covariances
}
\seealso{
\code{\link{stepreml_2way}} for generic version.
}
