% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepreml.r
\name{stepreml_2way}
\alias{stepreml_2way}
\title{calvin-dykstra solves for S1 < S2 < #3}
\usage{
stepreml_2way(
  M1,
  I1,
  M2,
  I2,
  M3,
  I3,
  max_iter = 50,
  err.tol = 1e-06,
  verbose = FALSE
)
}
\description{
calvin-dykstra solves for S1 < S2 < #3
}
