% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbalanced_EM.R
\name{sum_inverse}
\alias{sum_inverse}
\title{Compute (A + E/n)^(-1) for a vector of ns}
\usage{
sum_inverse(Sigma_E, Sigma_A, ns)
}
\arguments{
\item{ns}{A vector of doubles}

\item{A, E}{Symmetric nonnegative-definite square matrices}
}
\value{
A list of inverted matrices indexed by the vector \code{ns}
}
\description{
Compute (A + E/n)^(-1) for a vector of ns
}
