% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbalanced_EM.R
\name{svdsqrt}
\alias{svdsqrt}
\title{Use svd to compute a square root of the non-negative definite symmetric matrix A}
\usage{
svdsqrt(A)
}
\description{
Use svd to compute a square root of the non-negative definite symmetric matrix A
}
