% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoway.r
\name{twoway}
\alias{twoway}
\alias{manova_2way}
\alias{preml_2way}
\alias{stepreml_2way}
\title{Compute covariance estimates for half-sib balanced designs}
\usage{
manova_2way(object, ...)

preml_2way(object, ...)

stepreml_2way(object, ...)
}
\arguments{
\item{object}{An object for which a sum-of-squares matrix can be computed.}

\item{...}{Other arguments.}
}
\description{
Generic function sfor computing MANOVA estimates for a balanced half-sib design
using various algorithms
}
\seealso{
\code{\link{manova_2way_mat}}, \code{\link{preml_2way_mat}},
\code{\link{stepreml_2way_mat}} for computing estimates from sum-of-squares
matrices; \code{\link{manova_2way_df}}, \code{\link{preml_2way_df}},
\code{\link{stepreml_2way_df}} from a table of data.
}
