% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preml.R
\name{preml}
\alias{preml}
\alias{preml_1way}
\alias{preml_2way}
\alias{preml_2way_df}
\title{Compute 1- or 2-way pseudo-REML estimates}
\usage{
preml_1way(Mw, r, Mb, n)

preml_2way(M1, I1, M2, I2, M3, I3)

preml_2way_df(df, ...)
}
\arguments{
\item{Mw, Mb}{Sum-of-squares within and between matrices}

\item{r, n}{Number of}

\item{df}{Data frame for 2-way design}
}
\description{
Given either a tuple of sum-of-squares matrices or a dataframe, computes
preudo-REML estimates for the covariances
}
